extract logic to hook
