+++ src/pages/recipes/RecipeCreation.jsx
@@ -1,49 +1,20 @@
-import React, { useState, useMemo } from "react";
-import { Route, Routes, useNavigate } from "react-router";
+import React from "react";
+import { Route, Routes } from "react-router";
 import { Link } from "react-router-dom";
-import { v4 as uuidv4 } from "uuid";
 import { Text } from "../../components/Text/Text";
 import AddIngredient from "./AddIngredient";
 import { Card } from "../../components/Card/Card";
 import { Icon } from "../../components/Icon/Icon";
+import useIngredients from "../../hooks/useIngredients";
 
 const RecipeCreation = () => {
-  const [ingredients, setIngredients] = useState([]);
-  const [ingredientToEdit, setIngredientToEdit] = useState(null);
-  const navigate = useNavigate();
-
-  const editIngedient = (data) => {
-    setIngredients((previousIngredients) => {
-      const edited = previousIngredients.map((ingredient) =>
-        ingredient.id === data.id ? { ...data, id: data.id } : ingredient,
-      );
-
-      return edited;
-    });
-  };
-
-  const addIngredient = (data) => {
-    console.log(data);
-    if (ingredients.find(({ id }) => id === data.id)) {
-      editIngedient(data);
-      return;
-    }
-    setIngredients((previousIngredients) => [
-      ...previousIngredients,
-      { id: uuidv4(), ...data },
-    ]);
-    navigate(-1);
-  };
-
-  const deleteIngredient = (id) => {
-    setIngredients(ingredients.filter((ingredient) => ingredient.id !== id));
-    setIngredientToEdit(null);
-  };
-
-  const handleEditLinkClick = (data) => {
-    console.log(ingredients.find(({ id }) => id === data));
-    setIngredientToEdit(ingredients.find(({ id }) => id === data));
-  };
+  const {
+    ingredients,
+    ingredientToEdit,
+    addIngredient,
+    deleteIngredient,
+    handleEditLinkClick,
+  } = useIngredients();
 
   const RecipeCreationElement = () => (
     <form>
